// apps/web/src/generated/navigation.ts
// AUTO-GENERATED by @tentacrawl/cli — do not edit
import type { SidebarNavItem } from '@tentacrawl/ui';

import { LayoutDashboard, Target, Radar } from 'lucide-react';

export const navigationItems: SidebarNavItem[] = [
  {
    label: 'Dashboard',
    path: '/dashboard',
    icon: LayoutDashboard,
    order: 0,
  },
  {
    label: 'Page Scraper',
    path: '/scrape',
    icon: Target,
    order: 10,
  },
  {
    label: 'Website Crawler',
    path: '/crawl',
    icon: Radar,
    order: 20,
  },
];